Add more colors for members of our team
Monthly stats are shown with dedicated colors for the first 7 team members
from `activity.logins` variable.
Now there are 8 team members, so the last one in the list (fedinskiy)
receives a different random color for every graph on the same page.
This commit adds two additional colors, one for fedinskiy and another for a
potential future team member.